requirements.txt
asgiref==3.5.2; python_version >= '3.7'
certifi==2022.9.24; python_version >= '3.6'
charset-normalizer==2.1.1; python_version >= '3.6'
coreapi==2.3.3
coreschema==0.0.4
django==4.1.3
djangorestframework-simplejwt==5.2.2
djangorestframework==3.14.0
drf-yasg==1.21.4
gunicorn==20.1.0
idna==3.4; python_version >= '3.5'
inflection==0.5.1; python_version >= '3.5'
itypes==1.2.0
jinja2==3.1.2; python_version >= '3.7'
markupsafe==2.1.1; python_version >= '3.7'
packaging==21.3; python_version >= '3.6'
pyjwt==2.6.0; python_version >= '3.7'
pyparsing==3.0.9; python_full_version >= '3.6.8'
python-dotenv==0.21.0
pytz==2022.6
requests==2.28.1; python_version >= '3.7' and python_version < '4'
ruamel.yaml.clib==0.2.7; python_version < '3.11' and platform_python_implementation == 'CPython'
ruamel.yaml==0.17.21; python_version >= '3'
setuptools==65.6.0; python_version >= '3.7'
sqlparse==0.4.3; python_version >= '3.5'
uritemplate==4.1.1; python_version >= '3.6'
urllib3==1.26.12; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5' and python_version < '4'
